Having auth working with Hilt
