+++ app/src/main/java/com/lastaoutdoor/lasta/LastaApp.kt
@@ -1,0 +1,6 @@
+package com.lastaoutdoor.lasta
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp class LastaApp : Application()
