+++ app/src/main/java/com/lastaoutdoor/lasta/data/model/UserModel.kt
@@ -1,0 +1,3 @@
+package com.lastaoutdoor.lasta.data.model
+
+data class UserModel(val userId: String, val username: String?, val profilePictureUrl: String?)
