+++ app/src/main/java/com/lastaoutdoor/lasta/ui/MainActivity.kt
@@ -1,18 +1,21 @@
-package com.lastaoutdoor.lasta.view
+package com.lastaoutdoor.lasta.ui
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.lastaoutdoor.lasta.view.screen.LoginScreen
+import com.lastaoutdoor.lasta.ui.screen.LoginScreen
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class MainActivity : ComponentActivity() {
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
+    enableEdgeToEdge()
     setContent {
-
       // The navcontroller that allows to go from the login screen to the main menu screen
       val navController = rememberNavController()
       NavHost(navController = navController, startDestination = "login") {
