+++ app/src/main/java/com/lastaoutdoor/lasta/ui/screen/DiscoveryScreen.kt
@@ -1,0 +1,10 @@
+package com.lastaoutdoor.lasta.ui.screen
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun DiscoveryScreen() {
+  Column { Text("Discovery screen") }
+}
