+++ app/src/main/java/com/lastaoutdoor/lasta/ui/screen/ProfileScreen.kt
@@ -1,10 +1,10 @@
-package com.lastaoutdoor.lasta.view.screen
+package com.lastaoutdoor.lasta.ui.screen
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 
 @Composable
-fun DiscoveryScreen() {
-  Column { Text("Discovery screen") }
+fun ProfileScreen() {
+  Column { Text("Profile screen") }
 }
