+++ app/src/main/java/com/lastaoutdoor/lasta/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.lastaoutdoor.lasta.view.theme
+package com.lastaoutdoor.lasta.ui.theme
 
 import androidx.compose.ui.graphics.Color
