--- app/src/main/java/com/lastaoutdoor/lasta/view/screen/ProfileScreen.kt
@@ -1,10 +1,0 @@
-package com.lastaoutdoor.lasta.view.screen
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-
-@Composable
-fun ProfileScreen() {
-  Column { Text("Profile screen") }
-}
