+++ app/src/main/res/values/strings.xml
@@ -5,4 +5,5 @@
     <string name="tab_map">Map</string>
     <string name="tab_discover">Discover</string>
     <string name="tab_profile">Profile</string>
+    <string name="web_client_id">720666228102-mtt8qsp3fuljeatb9128qss6c2qf91t2.apps.googleusercontent.com</string>
 </resources>
