+++ build.gradle.kts
@@ -9,6 +9,7 @@ plugins {
     id("com.android.application") version "8.3.1" apply false
     id("org.jetbrains.kotlin.android") version "1.8.10" apply false
     id("com.google.gms.google-services") version "4.4.0" apply false
+    id("com.google.dagger.hilt.android") version "2.44" apply false
 }
 
 sonar {
